requirements.txt
langchain-ollama
langchain-groq
langchain-mistralai
mistralai
langchain_core
langchain-community
langchain-text-splitters
python-dotenv
pydantic
openai
langchain
pypdf
faiss-cpu
tiktoken
langchain_google_genai
streamlit
pytesseract 
pdf2image
google-cloud-vision
